+++ Core/Core/Courses/SmartSearch/View/SmartSearchFiltersView.swift
