+++ Core/Core/Courses/SmartSearch/View/SmartSearchHelpView.swift
